hw2 fix
