+++ models/contacts/index.js
@@ -1,0 +1,13 @@
+import { addContact } from "./addContact";
+import { getContactById } from "./getContactById";
+import { listContacts } from "./listContacts";
+import { removeContact } from "./removeContact";
+import { updateContact } from "./updateContact";
+
+export default {
+  addContact,
+  getContactById,
+  listContacts,
+  removeContact,
+  updateContact,
+};
