+++ models/contacts/listContacts.js
@@ -1,0 +1,5 @@
+import contacts from "../../db/contacts.json";
+
+export const listContacts = async () => {
+  return contacts;
+};
