+++ routes/api/contacts.js
@@ -1,43 +1,25 @@
 import express from "express";
-import model from "../../model/index";
-import { validation, validateUpdate, validateId } from "./validation";
+import {
+  getContacts,
+  getContactsById,
+  addContact,
+  removeContact,
+  udateContact,
+} from "../../controllers/contacts/index.js";
+import { validation } from "../../midllewares/validation/validation";
+import { validateId } from "../../midllewares/validation/validateId";
+import { validateUpdate } from "../../midllewares/validation/validateUpdate";
+
 const router = express.Router();
 
-router.get("/", async (req, res, next) => {
-  const contacts = await model.listContacts();
-  res.status(200).json(contacts);
-});
+router.get("/", getContacts);
 
-router.get("/:id", validateId, async (req, res, next) => {
-  const { id } = req.params;
-  const contact = await model.getContactById(id);
-  if (contact) {
-    return res.status(200).json(contact);
-  }
-  res.status(404).json({ message: "Not found" });
-});
+router.get("/:id", validateId, getContactsById);
 
-router.post("/", validation, async (req, res, next) => {
-  const newContact = await model.addContact(req.body);
-  res.status(201).json(newContact);
-});
+router.post("/", validation, addContact);
 
-router.delete("/:id", validateId, async (req, res, next) => {
-  const { id } = req.params;
-  const contact = await model.removeContact(id);
-  if (contact) {
-    return res.status(200).json({ message: "contact deleted" });
-  }
-  res.status(404).json({ message: "Not found" });
-});
+router.delete("/:id", validateId, removeContact);
 
-router.put("/:id", validateId, validateUpdate, async (req, res, next) => {
-  const { id } = req.params;
-  const contact = await model.updateContact(id, req.body);
-  if (contact) {
-    return res.status(200).json(contact);
-  }
-  res.status(404).json({ message: "Not found" });
-});
+router.put("/:id", validateId, validateUpdate, udateContact);
 
 export default router;
